UTIL download_maxmind_geo2ip_lite | move to php CLASS
